Cast print of numEntries in mod_webauthldap
lc->numEntries may be larger than an int.  Cast to unsigned long and
log with %lu.

Problem found by William Orr <will@worrbase.com>.

Reviewed-on: https://gerrit.stanford.edu/1019
Tested-by: Russ Allbery <rra@stanford.edu>

+++ modules/ldap/mod_webauthldap.c
@@ -711,8 +711,8 @@ webauthldap_dosearch(MWAL_LDAP_CTXT* lc)
             }
             if (lc->sconf->debug)
                 ap_log_error(APLOG_MARK, APLOG_INFO, 0, lc->r->server,
-                             "webauthldap(%s): search returned %d entries",
-                             lc->r->user, lc->numEntries);
+                             "webauthldap(%s): search returned %lu entries",
+                             lc->r->user, (unsigned long) lc->numEntries);
         }
         ldap_msgfree(res);
     }
